Добавлен компонент favorites-empty
